+++ src/olmo_core/nn/transformer/config.py
@@ -138,6 +138,11 @@ class TransformerBlockType(StrEnum):
     ➡️ :class:`FLABlock`
     """
 
+    default_scaled = "default_scaled"
+    """
+    ➡️ :class:`LayerNormScaledTransformerBlock` (applies LayerNorm Scaling)
+    """
+
 
 @dataclass
 class TransformerBlockConfig(Config):
@@ -182,6 +187,7 @@ class TransformerBlockConfig(Config):
     ) -> "TransformerBlockBase":
         from .block import (
             FLABlock,
+            LayerNormScaledTransformerBlock,
             MoEHybridReorderedNormTransformerBlock,
             MoEHybridTransformerBlock,
             MoEReorderedNormTransformerBlock,
@@ -204,6 +210,8 @@ class TransformerBlockConfig(Config):
         try:
             if self.name == TransformerBlockType.default:
                 return TransformerBlock(**kwargs)
+            elif self.name == TransformerBlockType.default_scaled:
+                return LayerNormScaledTransformerBlock(**kwargs)
             elif self.name == TransformerBlockType.reordered_norm:
                 return ReorderedNormTransformerBlock(**kwargs)
             elif self.name == TransformerBlockType.normalized: