+++ src/test/data/source_mixture_test.py
@@ -1,4 +1,5 @@
 import logging
+import math
 from itertools import chain
 from pathlib import Path
 
@@ -39,14 +40,17 @@ def test_source_mixture_config(tmp_path: Path, caplog, capsys):
     ]
 
     max_tokens = 5_000_000
+    sequence_length = 1024
+    global_batch_size = 1024 * 32
 
     config = SourceMixtureDatasetConfig(
         max_tokens=max_tokens,
         source_configs=source_configs,
         dtype=NumpyDatasetDType.uint32,
-        sequence_length=1024,
+        sequence_length=sequence_length,
         quiet=True,
         render_tables=True,
+        global_batch_size=global_batch_size,
     )
 
     # NOTE: We need to disable capsys so we can override log capture as
@@ -55,6 +59,14 @@ def test_source_mixture_config(tmp_path: Path, caplog, capsys):
         config.validate()
         mixture = config.build()
         assert isinstance(mixture, SourceMixtureDataset)
+
+        requested_instances = math.ceil(max_tokens / global_batch_size) * int(
+            global_batch_size / sequence_length
+        )
+        assert (
+            sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])
+            == requested_instances
+        ), f"Expected {requested_instances} instances, but got {sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])}"
         #  print(caplog.text)  # uncomment if you want to see the table
 
 
@@ -86,6 +98,7 @@ def test_dataset_mixture_config_validation():
         sequence_length=1024,
         quiet=True,
         render_tables=False,
+        global_batch_size=1024 * 32,
     )
     config.validate()
 
@@ -101,6 +114,7 @@ def test_dataset_mixture_config_validation():
         sequence_length=1024,
         quiet=True,
         render_tables=False,
+        global_batch_size=1024 * 32,
     )
 
     with pytest.raises(OLMoConfigurationError):
@@ -131,18 +145,29 @@ def test_dataset_mixture_build(tmp_path: Path):
     ]
 
     max_tokens = 5_000_000
+    sequence_length = 1024
+    global_batch_size = sequence_length * 32
 
     config = SourceMixtureDatasetConfig(
         max_tokens=max_tokens,
         source_configs=source_configs,
         dtype=NumpyDatasetDType.uint32,
-        sequence_length=1024,
+        sequence_length=sequence_length,
         quiet=True,
         render_tables=False,
+        global_batch_size=global_batch_size,
     )
 
     mixture = config.build()
     assert isinstance(mixture, SourceMixtureDataset)
+
+    requested_instances = math.ceil(max_tokens / global_batch_size) * int(
+        global_batch_size / sequence_length
+    )
+    assert (
+        sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])
+        == requested_instances
+    ), f"Expected {requested_instances} instances, but got {sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])}"
 
 
 def test_dataset_mixture_build_insufficient_source_data(tmp_path: Path):
@@ -176,6 +201,7 @@ def test_dataset_mixture_build_insufficient_source_data(tmp_path: Path):
         sequence_length=1024,
         quiet=True,
         render_tables=False,
+        global_batch_size=1024 * 32,
     )
 
     # Should raise exception because the target ratio for source 1 @50% (2.5M) is infeasible without repetition (default max_repetition_ratio=1)
@@ -213,14 +239,17 @@ def test_dataset_mixture_build_with_repetition(tmp_path: Path):
     ]
 
     max_tokens = 5_000_000
+    sequence_length = 1024
+    global_batch_size = sequence_length * 32
 
     config = SourceMixtureDatasetConfig(
         max_tokens=max_tokens,
         source_configs=source_configs,
         dtype=NumpyDatasetDType.uint32,
-        sequence_length=1024,
+        sequence_length=sequence_length,
         quiet=True,
         render_tables=False,
+        global_batch_size=global_batch_size,
     )
 
     mixture = config.build()
@@ -231,7 +260,16 @@ def test_dataset_mixture_build_with_repetition(tmp_path: Path):
 
     total_tokens = sum([item.tokens for item in all_paths])
     assert isinstance(mixture, SourceMixtureDataset)
-    assert total_tokens == 5_000_000
+
+    requested_instances = math.ceil(max_tokens / global_batch_size) * int(
+        global_batch_size / sequence_length
+    )
+    assert (
+        sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])
+        == requested_instances
+    ), f"Expected {requested_instances} instances, but got {sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])}"
+
+    assert total_tokens == 5013504
 
 
 def test_dataset_mixture_build_insufficient_source_max_fraction(tmp_path: Path):
@@ -269,6 +307,7 @@ def test_dataset_mixture_build_insufficient_source_max_fraction(tmp_path: Path):
         sequence_length=1024,
         quiet=True,
         render_tables=False,
+        global_batch_size=1024 * 32,
     )
 
     # Should raise exception because the target ratio for source 1 is infeasible because
@@ -305,14 +344,17 @@ def test_dataset_mixture_build_duplicate_paths(tmp_path: Path):
     ]
 
     max_tokens = 3_000_000
+    sequence_length = 1024
+    global_batch_size = sequence_length * 32
 
     config = SourceMixtureDatasetConfig(
         max_tokens=max_tokens,
         source_configs=source_configs,
         dtype=NumpyDatasetDType.uint32,
-        sequence_length=1024,
+        sequence_length=sequence_length,
         quiet=True,
         render_tables=False,
+        global_batch_size=global_batch_size,
     )
 
     expected = [str(sources["1"][0][0])] + [str(item[0]) for item in list(chain(*sources.values()))]
@@ -320,6 +362,13 @@ def test_dataset_mixture_build_duplicate_paths(tmp_path: Path):
     index = mixture.to_index()
     paths = mixture.to_paths()
     assert paths == expected
-    assert len(index) == 6
+    assert len(index) == 6, "Expected 6 unique paths in the index, but got {}".format(len(index))
     assert isinstance(mixture, SourceMixtureDataset)
-    assert len(mixture.sources) == 3
+    assert len(mixture.sources) == 3, "Expected 3 sources, but got {}".format(len(mixture.sources))
+    requested_instances = math.ceil(max_tokens / global_batch_size) * int(
+        global_batch_size / sequence_length
+    )
+    assert (
+        sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])
+        == requested_instances
+    ), f"Expected {requested_instances} instances, but got {sum([tokens // sequence_length for _, tokens in mixture.to_index().items()])}"